requirements.txt
streamlit>=1.38.0
pandas>=2.2.2
instructor>=1.4.2
pydantic>=2.9.1
openai>=1.45.0
langgraph>=0.2.21
pillow>=10.4.0